use CollectionLocation::CaseSensitivity enum

+++ core/libs/database/collection/collectionmanager_location.cpp
@@ -166,7 +166,7 @@ CollectionLocation CollectionManager::refreshLocation(const CollectionLocation& 
         access.db()->setAlbumRootType(location.id(),            type);
         access.db()->migrateAlbumRoot(location.id(),            identifier);
         access.db()->setAlbumRootPath(location.id(),            specificPath);
-        access.db()->setAlbumRootCaseSensitivity(location.id(), (int)CollectionLocation::UnknownCaseSensitivity);
+        access.db()->setAlbumRootCaseSensitivity(location.id(), CollectionLocation::UnknownCaseSensitivity);
 
         albumLoc->setLabel(label);
         albumLoc->identifier   = identifier;
@@ -203,7 +203,7 @@ CollectionLocation CollectionManager::refreshLocation(const CollectionLocation& 
         access.db()->setAlbumRootType(location.id(),            type);
         access.db()->setAlbumRootPath(location.id(),            QLatin1String("/"));
         access.db()->migrateAlbumRoot(location.id(),            d->volumeIdentifier(path));
-        access.db()->setAlbumRootCaseSensitivity(location.id(), (int)CollectionLocation::UnknownCaseSensitivity);
+        access.db()->setAlbumRootCaseSensitivity(location.id(), CollectionLocation::UnknownCaseSensitivity);
 
         albumLoc->setLabel(label);
         albumLoc->specificPath = QLatin1String("/");
@@ -897,7 +897,7 @@ void CollectionManager::updateLocations()
                 }
 
                 CoreDbAccess().db()->setAlbumRootCaseSensitivity(location->id(),
-                                                                 (int)location->caseSensitivity());
+                                                                 location->caseSensitivity());
             }
         }
